+++ spec/writer/html_index_writer_spec.rb
@@ -2,6 +2,7 @@ require './spec/helper'
 
 describe HtmlIndexWriter do
   let(:coverage_dir) { 'coverage' }
+  let(:config) { {theme: 'light', output_directory: 'coverage'} }
 
   after :each do
     FileUtils.remove_dir(coverage_dir) if File.exist?(coverage_dir)
@@ -9,7 +10,7 @@ describe HtmlIndexWriter do
 
   it "writes the index file" do
     FileUtils.makedirs(coverage_dir)
-    HtmlIndexWriter.new('content').write
+    HtmlIndexWriter.new('content', config).write
 
     expect(File).to exist('coverage/index.html')
   end